[MYDA-638][insu] Unit test: IS04 /insurances/car (#391)
* [MYDA-638][insu] Unit test: IS04 /insurances/car

새로운 템플릿 적용
기존 코드 수정
- insu: CarInsuranceMapper에 bigdeciaml 변환 누락
- insu: Dto에서 selfPayAmt를 BigDecimal로 변경
- insu: ExecutionTest.java 삭제

* remote와 충돌 해결
- InsuranceTestConstants 생성
- InsuranceContractServiceTest 주석 처리

+++ insu/src/main/java/com/banksalad/collectmydata/insu/car/dto/CarInsurance.java
@@ -9,6 +9,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -30,7 +31,7 @@ public class CarInsurance {
   private boolean ownDmgCoverage;
 
   private String selfPayRate;
-  private long selfPayAmt;
+  private BigDecimal selfPayAmt;
 
   /* additional fields */
   private String insuNum;
